Implement the My Orders page on the front end
The My Orders page allows users to submit a form that takes in an email
as input and emails the user, using Resend, a history of orders attached
to the provided email/customer. Form validation for this email
submission form was done with Zod and the body of the email was
formatted using react email components.